Separated endpoints into traits for easier editing Added missing `/forums/topics/{id}/posts` endpoint as `getForumTopicPosts($topicIc)`
